Add the Ratio script
Ratio reduces all the arguments by dividing each of them by their GCD.
It is useful to e.g. find out what screen resolution corresponds to
1920x1080 pixels.
(In order to find out that the answer is 16:9, call `Ratio 1920 1080'.
You will also learn that the scale factor, a.k.a. the GCD equals 120.)